+++ payment/src/main/java/com/goti/payment/controller/PaymentController.java
@@ -17,7 +17,6 @@ import com.goti.global.api.ApiSuccessResponse;
 import com.goti.payment.dto.request.PaymentRequest;
 import com.goti.payment.dto.response.PaymentResponse;
 import com.goti.payment.service.application.OrderPaymentService;
-import com.goti.payment.service.application.PaymentQueryService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -30,7 +29,6 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/api/v1/payments")
 public class PaymentController {
 	private final OrderPaymentService orderPaymentService;
-	private final PaymentQueryService paymentQueryService;
 
 	@Operation(
 		summary = "결제 요청",
@@ -61,6 +59,6 @@ public class PaymentController {
 		@PathVariable UUID orderId,
 		@AuthenticationPrincipal(expression = "id") UUID memberId
 	) {
-		return wrap(paymentQueryService.getByOrderId(orderId, memberId));
+		return wrap(orderPaymentService.getByOrderId(orderId, memberId));
 	}
 }
